+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- RariMe App => Rarimo App;
+
 ## [5.25.1] - 2025-10-16
 ### Added
 - ZK Liveness demo;